fix: solve problem with data displaying, change place of the logout-button
